+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+* Fix menu in row of `object-list-view` component.
+
 ## [0.7.0-beta.10] - 2016-10-06
 ### Changed
 * Updated dependency on `ember-flexberry-data` addon to v0.6.2-beta.13 version.